// packages/serve/src/client/api.ts
export type TreeNodeDataKind = typeof TreeNodeDataKind[keyof typeof TreeNodeDataKind];


export const TreeNodeDataKind = {
  action: 'action',
  sequence: 'sequence',
  selector: 'selector',
  parallel: 'parallel',
  ref: 'ref',
} as const;

export type TreeNodeDataStepsItemKind = typeof TreeNodeDataStepsItemKind[keyof typeof TreeNodeDataStepsItemKind];


export const TreeNodeDataStepsItemKind = {
  evaluate: 'evaluate',
  instruct: 'instruct',
} as const;

export type TreeNodeDataStepsItem = {
  kind: TreeNodeDataStepsItemKind;
  text: string;
};

export interface TreeNodeData {
  id: string;
  path: string;
  name: string;
  kind: TreeNodeDataKind;
  ref: string | null;
  steps?: TreeNodeDataStepsItem[];
  status: 'success' | 'failure' | 'running' | null;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  retries?: number;
  x: number;
  y: number;
}